refactored to move tests to conceptgraph package
